move tests to helper folder
